refactor: fold temporary error response factory into exception mapper
